Isolated nltk influences, and positioned their license file accordingly
