+++ mods/core/modules/algorithms/__init__.py
@@ -1,0 +1,1 @@
+import search # Load bearing __init__ file... as else the imports are not recognised
